+++ spring-ai-rag-back/src/main/java/com/lss/springairag/entity/KnowledgeChunk.java
@@ -14,7 +14,7 @@ public class KnowledgeChunk {
     @TableId(type = IdType.AUTO)
     private Long id;
 
-    private Integer fileId;
+    private Long fileId;
 
     private String documentId;
