+++ spring-ai-rag-back/src/main/java/com/lss/springairag/service/UserService.java
@@ -17,7 +17,7 @@ public interface UserService extends IService<User> {
 
     PageResult pageQuery(UserPageQueryDTO userPageQueryDTO);
 
-    void startOrStop(Integer status, Integer id);
+    void startOrStop(Integer status, Long id);
 
     void updateUser(UserDTO userDTO);
